requirements.txt
aiodns==3.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioresponses==0.7.8
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
attrs==25.3.0
authlib==1.6.0
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bandit==1.8.6
bcrypt==4.3.0
black==25.1.0
brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.9.2
cryptography==45.0.5
cyclopts==3.22.2
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
docutils==0.21.2
durationpy==0.10
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.3.0
execnet==2.1.1
fastapi==0.116.1
fastmcp==2.10.5
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
freezegun==1.5.3
frozenlist==1.7.0
fsspec==2025.5.1
ghp-import==2.1.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mccabe==0.7.0
mcp==1.11.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.2
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
mkdocs-section-index==0.3.10
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
mypy==1.16.1
mypy-extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
onnxruntime==1.22.1
opa-python-client==2.0.2
openai==1.95.1
openapi-pydantic==0.5.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-prometheus==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.4.0
pre-commit==4.2.0
prometheus-client==0.22.1
propcache==0.3.2
protobuf==6.31.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pycares==4.9.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyflakes==3.4.0
pygments==2.19.2
pyjwt==2.10.1
pymdown-extensions==10.16
pyperclip==1.9.0
pypika==0.48.9
pyproject-hooks==1.2.0
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-benchmark==5.1.0
pytest-cov==6.2.1
pytest-httpx==0.35.0
pytest-mock==3.14.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyyaml==6.0.2
pyyaml-env-tag==1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.1
stevedore==5.4.1
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
